Move /visualise to /y/visualise
Following the changes from from pull request 3126, the /visualise path
was affected and has been returning 404.

As a result of the aforementioned PR, this commit changes the route for
the visualise tool from /visualise to /y/visualise.

This is being done as the smaller change, compared to the other option
which would entail other changes like defining, allocating and storing
each smart answer's visualise content id/base path on the content store

We have decided to go with this option to make visualise available to
content designers under the /y (i.e /y/visualise) and also allow smart
answer controller to respond independent of a content store entry.